AUTOSoluton and SLPoint are very tightly coupled, move them into one module
